+++ Arm_IK_Sim_Unity/Assets/Scripts/CopyTransform.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CopyTransform : MonoBehaviour
+{
+    [SerializeField] private Transform target;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        transform.position = target.position;
+        transform.rotation = target.rotation;
+    }
+}
