+++ Arm_IK_Sim_Unity/Assets/Scripts/CopyTransform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 35afa62f336741a4ca53e37c2188d941
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
